feat: Add mark all as read functionality to pop-up menu
This commit adds the functionality to mark all messages as read in the pop-up menu of the Messages screen. When the "Mark all as read" option is tapped, the code calls the `markAllAsRead()` method from the `ApiService` class to send a request to the server and update the read status of all messages. If the request is successful, a success message is displayed, otherwise, an error is handled.still needs handeling tho in shared prefs