refactor: Update User.fromJson to handle dynamic json input
The User.fromJson method has been refactored to handle dynamic json input. It now checks the type of the json parameter and creates a User object accordingly. If the json is a Map<String, dynamic>, it extracts the id and username values. If the json is a String, it sets the id as the json value and sets the username as 'you'. If the json is of any other type, an exception is thrown. This update improves the flexibility and robustness of the User.fromJson method.